+++ packages/sanity/src/ui/changes.md
@@ -1,12 +1,49 @@
 ## SpacerButton
+
+- Moved to desk/components
 - Used by
-    -  desk: 
-- Moved to  desk/components
+  - desk:
+    - Banner
+    - DocumentStatusBar
 
 ## DocumentStatusIndicator
-- Used by:
-    - core: ReferenceInput
-    - core: SearchResultItemPreview
-    - desk: PaneItemPreview
-    - desk: DocumentStatusLine
+
 - Moved to: core/components
+- Used by:
+  - core:
+    - ReferenceInput
+    - SearchResultItemPreview
+    - PaneItemPreview
+    - DocumentStatusLine
+
+## LoadingBlock
+
+- Moved to: core/components
+- Used by:
+  - core:
+    - ReferenceItem
+    - GridItem
+    - PreviewItem
+    - ImagePreview
+    - ImageToolInput
+    - ImageTool
+    - AssetSource/\*
+    - LocaleProvider
+    - createLoginComponent
+    - AuthBoundary
+    - StudioLayout
+    - StudioProvider
+    - NewDocumentList
+    - ResourcesMenuItems
+    - ImageReferencePreview
+    - WorkspaceAuth
+  - comments:
+    - CommentsListStatus
+    - MentionsMenu
+  - desk:
+    - ConfirmDeleteDialog
+    - FormView
+    - ChangesInspector
+    - Timeline
+    - DocumentListPaneContent
+    - LoadingPane